system: chunky boi gets all the things
At a quick standup meeting today we realized that the system container
is perfect for staging & development of nftables rules and ntp.conf
before deploying with the smaller application containers.

So, the system container will from now on be the superset of all the
smaller application containers, and it will be optional for the latter
to include any BusyBox tools.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>