configs: refactor nftables container, add shell + basic tooling
This container is primarily intended as a reference with Infix, where it
comes bundled by default.  So there's an implied requirement to keep the
container image as small as possible.

However, the container must strike a balance between size and usability.
To that end a basic shell must be available and the smallest is BusyBox,
which also means it's easier to just drop tini in place of BusyBox init.

The total size increases from 542 kB to 692 kB (+150 kB), a small price
for the added functionality (ash, vi, ls/cat/mv, unshare/nsenter, etc).
As a spin-off the added tooling allows for using the container as base
for regression tests in Infix \o/

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>