configs: add image description to each defconfig
This was a tricky one, beacuse there's a bug in /buildroot/fs/oci/oci.mk that
cause any value with spaces to be split regardless of quoting or escaping. As
a workaround I've used U+2800 (Braille Pattern Blank) which survives this,
but it is not possible to edit with menuconfig.

The problem is $(foreach label,$(OCI_LABELS),...), which splits on all and any
whitespace, regardless of quoting.  It doesn't understand shell quotes or
escaping. So the following:

    .url=... .title=... .description='text with spaces'

Gets split into multiple items at every space, not just at the dots between
labels.  Compared to how ENTRYPOINT/CMD are handled:

    $(shell eval printf -- "--entrypoint\ \'%s\'\ " $(BR2_TARGET_ROOTFS_OCI_ENTRYPOINT))

This properly handles spaces with shell quoting.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>